Update display-projection-mode extension
